Addressing review comments.

+++ koku/masu/processor/ocp/ocp_report_charge_updater.py
@@ -187,7 +187,7 @@ class OCPReportChargeUpdater(OCPCloudUpdaterBase):
             None
 
         """
-        aws_markup_value = 0.0
+        aws_markup_value = Decimal(0.0)
         infra_map = self.get_infra_map()
         infra_tuple = infra_map.get(self._provider_uuid)
         cluster_id = get_cluster_id_from_provider(self._provider_uuid)
@@ -196,11 +196,11 @@ class OCPReportChargeUpdater(OCPCloudUpdaterBase):
             with CostModelDBAccessor(self._schema, aws_uuid,
                                      self._column_map) as cost_model_accessor:
                 markup = cost_model_accessor.get_markup()
-                aws_markup_value = float(markup.get('value', 0)) / 100
+                aws_markup_value = Decimal(markup.get('value', 0) / 100)
         with CostModelDBAccessor(self._schema, self._provider_uuid,
                                  self._column_map) as cost_model_accessor:
             markup = cost_model_accessor.get_markup()
-            ocp_markup_value = float(markup.get('value', 0)) / 100
+            ocp_markup_value = Decimal(markup.get('value', 0) / 100)
 
         with OCPReportDBAccessor(self._schema, self._column_map) as accessor:
             LOG.info('Updating OpenShift markup for'
